moved make_address to test.utils
