+++ raiden/tests/smart_contracts/netting_channel/test_updatetransfer.py
@@ -4,8 +4,9 @@ from ethereum.tester import TransactionFailed
 from coincurve import PrivateKey
 
 from raiden.messages import DirectTransfer
-from raiden.utils import privatekey_to_address, sha3, make_address
+from raiden.utils import privatekey_to_address, sha3
 from raiden.tests.utils.transfer import make_direct_transfer_from_channel
+from raiden.tests.utils.factories import make_address
 
 
 def test_transfer_update_event(tester_state, tester_channels, tester_events):
